+++ expr/func_persecond.go
@@ -59,12 +59,14 @@ func (s *FuncPerSecond) Exec(cache map[Req][]models.Series) ([]models.Series, er
 			}
 		}
 		s := models.Series{
-			Target:     fmt.Sprintf("perSecond(%s)", serie.Target),
-			QueryPatt:  fmt.Sprintf("perSecond(%s)", serie.QueryPatt),
-			Tags:       serie.Tags,
-			Datapoints: out,
-			Interval:   serie.Interval,
-			Meta:       serie.Meta,
+			Target:       fmt.Sprintf("perSecond(%s)", serie.Target),
+			QueryPatt:    fmt.Sprintf("perSecond(%s)", serie.QueryPatt),
+			Tags:         serie.Tags,
+			Datapoints:   out,
+			Interval:     serie.Interval,
+			Meta:         serie.Meta,
+			QueryMDP:     serie.QueryMDP,
+			QueryPNGroup: serie.QueryPNGroup,
 		}
 		outputs = append(outputs, s)
 		cache[Req{}] = append(cache[Req{}], s)