+++ expr/tagquery/query.go
@@ -20,6 +20,9 @@ func init() {
 	QueryHash = util.NewFnv32aStringWriter
 }
 
+//Query is a set of graphite tag expressions as used by the graphite `seriesByTag` function.
+//tag expression definitions: https://graphite.readthedocs.io/en/latest/tags.html#querying
+//seriesByTag documentation: https://graphite.readthedocs.io/en/latest/functions.html#graphite.render.functions.seriesByTag
 type Query struct {
 	// clause that operates on LastUpdate field
 	From int64
@@ -33,6 +36,10 @@ type Query struct {
 	tagClause int
 }
 
+//NewQueryFromStrings parses a list of graphite tag expressions as used by the graphite `seriesByTag` function.
+//tag expression definitions: https://graphite.readthedocs.io/en/latest/tags.html#querying
+//seriesByTag documentation: https://graphite.readthedocs.io/en/latest/functions.html#graphite.render.functions.seriesByTag
+//Some possible tag expressions are: "status=200", "path!=/", "name=~cpu\..*" (`name` is  a special tag which is automatically applied to the metric name).
 func NewQueryFromStrings(expressionStrs []string, from int64) (Query, error) {
 	var res Query
 	expressions, err := ParseExpressions(expressionStrs)